Issues #8 and #9 improve docker image names and tags
This is a speculative PR to start continue the discussion of how we want the
image names and tags to be.

This now builds an image with the following tags (currently with SNAPSHOT):
 - openjdk:8
 - openjdk:8u102
 - openjdk:8u102-1-SNAPSHOT
 - openjdk:8u102-1-SNAPSHOT-2016-09-15_01_31
 - openjdk:8u102-2016-09-15_01_31
 - openjdk:latest

This allows a dependent image to use:

 - the latest version of openjdk
 - the latest version of openjdk:8
 - the latest build of a specific version of openjdk:8u102
 - a specific build of a specific version of openjdk:8u102-1-SNAPSHOT by version
 - a specific build of a specific version of openjdk:8u102-2016-09-15_01_31 by timestamp
 - a specific build of a specific version of openjdk:8u102-1-SNAPSHOT-2016-09-15_01_31 by version and timestamp

The version is now just a single digit release number and if future we could expect multiple
versions of java to be built under a single release number:

 - openjdk:8u108-42-2017-02-01_12_00
 - openjdk:8u108-42
 - openjdk:8u108
 - openjdk:8
 - openjdk:9u2-42-2017-02-01_12_01
 - openjdk:9u2-42
 - openjdk:9u2
 - openjdk:9

It may be a bit over the top having both release number and timestamp, but tags are cheap!